implemented suggested changes
